Update giao dien form login
Chỉnh sửa jpanel thành frame
