+++ nbproject/project.properties
@@ -72,7 +72,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=PRO1014_QUAN_LY_THU_VIEN
-main.class=pro1014_quan_ly_thu_vien.PRO1014_QUAN_LY_THU_VIEN
+main.class=views.JFrame_Login
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
